+++ src/controllers/highlights/HighlightPhrase.cpp
@@ -19,18 +19,19 @@ bool HighlightPhrase::operator==(const HighlightPhrase &other) const
     return std::tie(this->pattern_, this->showInMentions_, this->hasSound_,
                     this->hasAlert_, this->isRegex_, this->isCaseSensitive_,
                     this->soundUrl_, this->color_, this->globalHighlight_,
-                    this->channels_) ==
+                    this->channels_, this->ExcludedChannels_) ==
            std::tie(other.pattern_, other.showInMentions_, other.hasSound_,
                     other.hasAlert_, other.isRegex_, other.isCaseSensitive_,
                     other.soundUrl_, other.color_, other.globalHighlight_,
-                    other.channels_);
+                    other.channels_, other.ExcludedChannels_);
 }
 
 HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                  bool hasAlert, bool hasSound, bool isRegex,
                                  bool isCaseSensitive, const QString &soundUrl,
                                  QColor color, bool globalHighlight,
-                                 std::vector<std::string> channels)
+                                 std::vector<std::string> channels,
+                                 std::vector<std::string> ExcludedChannels)
     : pattern_(pattern)
     , showInMentions_(showInMentions)
     , hasAlert_(hasAlert)
@@ -47,6 +48,7 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                    : QRegularExpression::CaseInsensitiveOption))
     , globalHighlight_(globalHighlight)
     , channels_(channels)
+    , ExcludedChannels_(ExcludedChannels)
 {
     this->color_ = std::make_shared<QColor>(color);
 }
@@ -56,7 +58,8 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                  bool isCaseSensitive, const QString &soundUrl,
                                  std::shared_ptr<QColor> color,
                                  bool globalHighlight,
-                                 std::vector<std::string> channels)
+                                 std::vector<std::string> channels,
+                                 std::vector<std::string> ExcludedChannels)
     : pattern_(pattern)
     , showInMentions_(showInMentions)
     , hasAlert_(hasAlert)
@@ -74,6 +77,7 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                    : QRegularExpression::CaseInsensitiveOption))
     , globalHighlight_(globalHighlight)
     , channels_(channels)
+    , ExcludedChannels_(ExcludedChannels)
 {
 }
 
@@ -140,6 +144,11 @@ bool HighlightPhrase::isGlobalHighlight() const
 const std::vector<std::string> &HighlightPhrase::getChannels() const
 {
     return this->channels_;
+}
+
+const std::vector<std::string> &HighlightPhrase::getExcludedChannels() const
+{
+    return this->ExcludedChannels_;
 }
 
 }  // namespace chatterino